+++ algorithm/convolution_int.cpp
@@ -10,7 +10,7 @@
 
 template <class T, std::enable_if_t<std::is_integral_v<T>>* = nullptr>
 std::vector<T> convolution_int(const std::vector<T>& a, const std::vector<T>& b) {
-    const usize n = a.size(), m = b.size();
+    const int n = a.size(), m = b.size();
     if (n == 0 || m == 0) return {};
     static constexpr u64 MOD1 = 754974721, MOD2 = 167772161, MOD3 = 469762049;
     static constexpr u64 M2M3 = MOD2 * MOD3, M1M3 = MOD1 * MOD3, M1M2 = MOD1 * MOD2;
@@ -23,7 +23,7 @@ std::vector<T> convolution_int(const std::vector<T>& a, const std::vector<T>& b)
     std::vector<T> c2 = convolution_mod<T, MOD2>(a, b);
     std::vector<T> c3 = convolution_mod<T, MOD3>(a, b);
     std::vector<T> c(n + m - 1);
-    for (const usize i : rep(0, n + m - 1)) {
+    for (const int i : rep(0, n + m - 1)) {
         u64 x = 0;
         x += (c1[i] * I1) % MOD1 * M2M3;
         x += (c2[i] * I2) % MOD2 * M1M3;
