+++ math/factorize.cpp
@@ -4,9 +4,9 @@
 #include <vector>
 #include "../utility/int_alias.cpp"
 
-template <class T> std::vector<std::pair<T, usize>> factorize(T x) {
+template <class T> std::vector<std::pair<T, int>> factorize(T x) {
     assert(x > 0);
-    std::vector<std::pair<T, usize>> ret;
+    std::vector<std::pair<T, int>> ret;
     for (T i = 2; i * i <= x; ++i) {
         if (x % i == 0) {
             ret.emplace_back(i, 0);
